requirements.txt
absl-py==2.3.1
aiobotocore==2.24.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aioitertools==0.12.0
aiosignal==1.4.0
alembic==1.16.4
amqp==5.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
appdirs==1.4.4
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
asyncssh==2.21.0
atpublic==6.0.1
attrs==25.3.0
Automat==25.4.16
aws-hadoop==1.0.dev2
babel==2.17.0
backrefs==5.9
bcrypt==4.3.0
beautifulsoup4==4.13.4
billiard==4.2.1
black==23.12.1
bleach==6.2.0
blinker==1.9.0
blis==1.2.1
bokeh==3.7.3
boto3==1.39.16
botocore==1.39.11
cachetools==5.5.2
catalogue==2.0.10
celery==5.5.3
certifi==2025.7.14
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpathlib==0.21.1
cloudpickle==3.1.1
cmdstanpy==1.2.5
colorama==0.4.6
colorlog==6.9.0
comm==0.2.3
confection==0.1.5
configobj==5.0.9
constantly==23.10.4
contourpy==1.3.3
coverage==7.10.1
cryptography==45.0.5
cssselect==1.3.0
cycler==0.12.1
cymem==2.0.11
Cython==3.1.2
dask==2023.12.1
databricks-sdk==0.60.0
debugpy==1.8.15
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
dictdiffer==0.9.0
diskcache==5.6.3
distlib==0.4.0
distributed==2023.12.1
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
docker-pycreds==0.4.0
dpath==2.2.0
dulwich==0.24.1
dvc==3.63.0
dvc-data==3.16.12
dvc-http==2.32.0
dvc-objects==5.1.1
dvc-render==1.0.2
dvc-studio-client==0.22.0
dvc-task==0.40.2
dvclive==3.48.5
entrypoints==0.4
executing==2.2.0
fabric==3.2.2
fastapi==0.116.1
fastjsonschema==2.21.1
filelock==3.18.0
flake8==6.1.0
Flask==3.1.1
flatbuffers==25.2.10
flatten-dict==0.4.2
flufl.lock==8.2.0
fonttools==4.59.0
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2025.7.0
funcy==2.0
gast==0.6.0
ghp-import==2.1.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.40.3
google-pasta==0.2.0
grandalf==0.8
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.2.3
grpcio==1.74.0
gto==1.8.0
h11==0.16.0
h5py==3.14.0
holidays==0.77
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.34.3
hydra-core==1.3.2
hyperlink==21.0.0
identify==2.6.12
idna==3.10
imageio==2.37.0
imbalanced-learn==0.14.0
imblearn==0.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
incremental==24.7.2
iniconfig==2.1.0
invoke==2.2.0
ipykernel==6.30.0
ipython==9.4.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
isort==5.13.2
itemadapter==0.12.0
itemloaders==1.3.2
iterative-telemetry==0.0.10
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.6
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
keras==3.11.0
kiwisolver==1.4.8
kombu==5.5.4
langcodes==3.5.0
language_data==1.3.0
lark==1.2.2
lazy_loader==0.4
libclang==18.1.1
lightgbm==4.6.0
lime==0.2.0.1
llvmlite==0.44.0
locket==1.0.0
loguru==0.7.3
lxml==6.0.0
lz4==4.4.4
Mako==1.3.10
marisa-trie==1.2.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mergedeep==1.3.4
mistune==3.1.3
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
mkdocs-material==9.6.16
mkdocs-material-extensions==1.3.1
ml_dtypes==0.5.3
mlflow==2.22.1
mlflow-skinny==2.22.1
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.4
murmurhash==1.0.13
mypy==1.17.0
mypy_extensions==1.1.0
mysql-connector-python==9.4.0
namex==0.1.0
narwhals==2.0.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.1
nodeenv==1.9.1
notebook==7.4.4
notebook_shim==0.2.4
numba==0.61.2
numpy==1.26.4
nvidia-ml-py==13.580.82
omegaconf==2.3.0
opencv-python==4.11.0.86
opentelemetry-api==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
opt_einsum==3.4.0
optree==0.17.0
optuna==3.6.2
orjson==3.11.3
overrides==7.7.0
packaging==24.2
paginate==0.5.7
pandas==2.3.1
pandera==0.25.0
pandocfilters==1.5.1
paramiko==4.0.0
parsel==1.10.0
parso==0.8.4
partd==1.4.2
pathspec==0.12.1
patsy==1.0.1
pep8==1.7.1
pillow==10.4.0
platformdirs==4.3.8
plotly==5.24.1
pluggy==1.6.0
pmdarima==2.0.4
pre-commit==3.8.0
preshed==3.0.10
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.2
prophet==1.1.7
Protego==0.5.0
protobuf==5.29.5
psutil==7.0.0
psycopg2-binary==2.9.10
pure_eval==0.2.3
py4j==0.10.9.7
pyarrow==19.0.1
pyarrow-hotfix==0.7
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.11.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyDispatcher==2.0.7
pydot==4.0.1
pyflakes==3.1.0
pygit2==1.18.2
Pygments==2.19.2
pygtrie==2.5.0
pymdown-extensions==10.16.1
pymongo==4.13.2
PyMySQL==1.1.2
PyNaCl==1.5.0
pynvml==13.0.1
pyOpenSSL==25.1.0
pyparsing==3.2.3
pypi-publisher==0.0.4
pyspark==3.5.6
pytest==7.4.4
pytest-cov==4.1.0
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
python-multipart==0.0.20
python-terraform==0.10.1
pytz==2025.2
pywin32==311
pywinpty==2.0.15
PyYAML==6.0.2
pyyaml_env_tag==1.1
pyzmq==27.0.0
queuelib==1.8.0
referencing==0.36.2
regex==2025.7.31
requests==2.32.4
requests-file==2.1.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.1.0
rpds-py==0.26.0
rsa==4.9.1
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.12
s3fs==2025.7.0
s3transfer==0.13.1
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.7.1
scipy==1.16.1
scmrepo==3.5.2
Scrapy==2.13.3
seaborn==0.13.2
semver==3.0.4
Send2Trash==1.8.3
sentry-sdk==1.45.1
service-identity==24.2.0
setproctitle==1.3.6
shap==0.48.0
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.2
six==1.17.0
slicer==0.0.8
smart_open==7.3.0.post1
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.42
sqlparse==0.5.3
sqltrie==0.11.2
srsly==2.5.1
stack-data==0.6.3
stanio==0.5.1
starlette==0.47.2
statsmodels==0.14.5
sympy==1.14.0
tabulate==0.9.0
tblib==3.1.0
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
terminado==0.18.1
thinc==8.3.4
threadpoolctl==3.6.0
tifffile==2025.6.11
tinycss2==1.4.0
tldextract==5.3.0
tokenizers==0.21.4
tomlkit==0.13.3
toolz==1.0.0
torch==2.7.1
torchvision==0.22.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.54.1
Twisted==25.5.0
typeguard==4.4.4
typer==0.16.0
types-python-dateutil==2.9.0.20250708
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
uvicorn==0.35.0
vine==5.1.0
virtualenv==20.32.0
voluptuous==0.15.2
w3lib==2.3.1
waitress==3.0.2
wandb==0.18.3
wasabi==1.1.3
watchdog==6.0.0
watchfiles==1.1.0
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
widgetsnbextension==4.0.14
win32_setctime==1.2.0
wrapt==1.17.2
xgboost==1.7.6
xyzservices==2025.4.0
yarl==1.20.1
zc.lockfile==3.0.post1
zict==3.0.0
zipp==3.23.0
zope.interface==7.2
